Import ox-reveal.
Hack. ox-reveal has a version incompatability with newer versions of Org
mode so can't be installed automatically. There's work underway to fix
this but it's not ready in time.

This is a fork with some changes to add styles for numbering, badges for
language types, etc. This will get removed and folded in to the main
branch later.