Update the MVC examples, slides, to pass items to View.update().
The previous version of the code required the View to fetch the items
from the Model, with the Model providing a getter to make the code
simpler.

This complicated the explanation significantly, and also breaks the
idea of the data flowing one-way around the MVC ring.

This code keeps the View's update() method, but now the Model calls
this method at appropriate times, passing in a copy of the current
list of items.