Apply some CSS rules to only the notes
The HTML_HEAD entry in settings.org applies to both the slides and the
notes export, which means they share html-export.css.

This causes problems when we need different styles for both.

As a quick fix, apply a "preso" class to the body when exporting slides,
and change some of style in html-export.css to only apply to body
elements that do not have this class.